requirements.txt
discord.py~=2.5.1
python-dotenv~=1.0.1
aiosqlite~=0.21.0